fix(ledger): signed reward-reserves delta (#796) + MIR no-panic solvency (#803)
#796 delta_reserves is now signed i128 end-to-end. Haskell applyRUpd uses
  addDeltaCoin over a signed DeltaCoin, so reserves INCREASE at a reward update
  when epoch_fees > treasury_cut + total_distributed (degraded/stalled epoch);
  dugite's u64 + saturating_sub silently dropped the reserves-credit leg. The 4
  compute sites now use unsaturated signed arithmetic; a shared
  apply_reserves_delta(u64, i128) helper replaces the 7 inlined apply sites with
  a sign branch (checked_sub when d>=0, checked_add(-d) when d<0). reward_debug
  dump field widened to i128. SNAPSHOT_VERSION 26->27 (PendingRewardUpdate is
  serialized into the snapshot). Test test_reward_zero_reserves_no_expansion,
  which pinned the bug, now asserts the correct +800k reserves credit.

#803 apply_pending_mir rewritten to mirror Haskell Rules.Mir.applyMIR instead of
  panicking on adversarial underflow. Drains all four pending accumulators up
  front (std::mem::take — so dsIRewards is wiped in BOTH branches, matching
  Haskell), filters to registered credentials, folds the pot-to-pot deltas into
  availability (available_reserves = reserves - delta_reserves + delta_treasury,
  the two-non-negative-magnitude encoding of Haskell's single signed deltaReserves
  — oracle-confirmed no cross-term), then does the single combined all-or-nothing
  solvency gate totR<=availableReserves && totT<=availableTreasury (plus a
  non-negative per-credential guard as defense-in-depth for the documented Phase-1
  MIR gaps). On failure: NoMirTransfer — pots byte-identical, pending cleared,
  warn, no panic. Byte-exact on all valid history; only the adversarial/insolvent
  path changes from crash to no-op. MIR is pre-Conway-only (not live-reachable).

cargo nextest --workspace: 7179 passed; clippy/fmt/workspace-build clean; SNAPSHOT_VERSION=27.

+++ crates/dugite-ledger/src/state/reward_debug.rs
@@ -45,7 +45,9 @@ pub struct Scalars {
     pub treasury_pre_rupd: u64,
     pub ss_fee: u64,
     pub bprev_total_blocks: u64,
-    pub delta_reserves: u64,
+    /// Signed RUPD reserves adjustment (see `PendingRewardUpdate::delta_reserves`
+    /// and issue #796) — negative means reserves increased this boundary.
+    pub delta_reserves: i128,
     pub delta_treasury: u64,
     pub total_rupd_credits: u64,
     /// Decentralisation `d` captured from the previous epoch — exact
